Extract Auth test in separate files
